move common speed formatting code to util.rs
